[v0.91.2][maintainability] Split adl/src/gws_live_content_card_roundtrip.rs by responsibility (Closes #3154)
